pyproject.toml
[tool.poetry]
name = "journal-bot-backend"
version = "0.1.0"
description = ""
authors = ["nwilson314 <nsw.wilson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.2"}
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
llama-index = "^0.11.18"
sqlalchemy = "^2.0.36"
sqlmodel = "^0.0.22"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
